+++ www/go/base/util/StringHelper.php
@@ -1246,7 +1246,7 @@ class StringHelper {
 			}
 		}
 
-		mt_srand((int) microtime() * 1000000);
+		mt_srand((int) microtime(true) * 1000000);
 
 		// Generate array of allowed characters by removing disallowed
 		// characters from array.