+++ www/go/base/util/StringHelper.php
@@ -1219,6 +1219,7 @@ class StringHelper {
 		// Generate array of allowable characters.
 		$characters_allow = explode(',', $characters_allow);
 
+		$array_allow = [];
 		for ($i = 0; $i < count($characters_allow); $i ++) {
 			if (substr_count($characters_allow[$i], '-') > 0) {
 				$character_range = explode('-', $characters_allow[$i]);
@@ -1233,7 +1234,7 @@ class StringHelper {
 
 		// Generate array of disallowed characters.
 		$characters_disallow = explode(',', $characters_disallow);
-
+		$array_disallow = [];
 		for ($i = 0; $i < count($characters_disallow); $i ++) {
 			if (substr_count($characters_disallow[$i], '-') > 0) {
 				$character_range = explode('-', $characters_disallow[$i]);
@@ -1246,7 +1247,7 @@ class StringHelper {
 			}
 		}
 
-		mt_srand(( double ) microtime() * 1000000);
+		mt_srand((int) microtime(true) * 1000000);
 
 		// Generate array of allowed characters by removing disallowed
 		// characters from array.