+++ www/go/core/orm/Entity.php
@@ -188,13 +188,17 @@ abstract class Entity extends Property {
 	 * $models = ModelWithDoublePK::findById("1-1");
 	 * ```
 	 *
-	 * @param string $id
+	 * @param string|null $id
 	 * @param string[] $properties
 	 * @param bool $readOnly
 	 * @return ?static
 	 */
-	public static final function findById(string $id, array $properties = [], bool $readOnly = false): ?Entity
+	public static final function findById(?string $id, array $properties = [], bool $readOnly = false): ?Entity
 	{
+		if($id == null) {
+			return null;
+		}
+
 		return static::internalFindById($id, $properties, $readOnly);
 	}
 
@@ -204,13 +208,11 @@ abstract class Entity extends Property {
 	 * Check if an ID exists in the database in the most efficient way. It also caches the result
 	 * during the same request.
 	 *
-	 * @param $id
+	 * @param string|null $id
 	 * @return bool
-	 * @throws Exception
 	 */
-	public static function exists($id): bool
+	public static function exists(?string $id): bool
 	{
-
 		if(empty($id)) {
 			return false;
 		}