+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Core: export to web page (html)
+
 24-11-2022 6.6.135
 - Calendar: tasklist_id cannot be null error on user create.