+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - Core: export to web page (html)
 - Fixed: dav_event sql data column too small #916
+- Fixed: core_auth_remember_me foreign key error #913
 
 24-11-2022 6.6.135
 - Calendar: tasklist_id cannot be null error on user create.