+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Fixed: core_auth_remember_me foreign key error #913
 - Fixed: Addressbook CSV import bug #902
 - Fixed: Semaphore buildup due to faulty unlock #901
+- Fixed: 6.6.128 - PHP8.1 complains about loosing precision #907
+
 
 24-11-2022 6.6.135
 - Calendar: tasklist_id cannot be null error on user create.