+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fixed: Addressbook CSV import bug #902
 - Fixed: Semaphore buildup due to faulty unlock #901
 - Fixed: 6.6.128 - PHP8.1 complains about loosing precision #907
+- Jitsi: Add JWT capability to jitsimeet module #900
 
 
 24-11-2022 6.6.135