+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+29-11-2022 6.6.136
 - Core: export to web page (html)
 - Fixed: dav_event sql data column too small #916
 - Fixed: core_auth_remember_me foreign key error #913